comprendre-logement/requirements.txt
streamlit
google-generativeai
chromadb
pysqlite3-binary
sentence-transformers